fix: footer on mobile
